eval-app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amazon-textract-caller==0.2.4
amazon-textract-response-parser==1.0.3
amazon-textract-textractor==1.8.4
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.35.59
botocore==1.35.59
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
editdistance==0.8.1
entrypoints==0.4
executing==2.1.0
faicons==0.2.2
fastjsonschema==2.20.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
htmltools==0.6.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
humanize==4.11.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-ollama==0.2.0
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langfuse==2.54.1
langsmith==0.1.142
linkify-it-py==2.0.3
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.5
nbclassic==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
ollama==0.3.3
openai==1.54.4
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pins==0.8.6
pip==24.2
platformdirs==4.3.6
plotly==5.24.1
posit-sdk==0.6.0
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
rsconnect==0.1.3
rsconnect-jupyter==1.8.1
rsconnect_python==1.24.0
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
semver==2.13.0
Send2Trash==1.8.3
setuptools==75.1.0
shiny==1.2.1
shinywidgets==0.3.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.2
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
